requirements.txt
streamlit
python-dotenv
langchain-core
langchain-community
langchain-groq
scikit-learn
fpdf2
mysql-connector-python
google-generativeai
PyPDF2